Require rdflib 4.0
RDFLib 4.0 was released over 2 years ago. This makes Surf usable with
modern versions of RDFLib, and allows removing some old compatibility
code.

We no longer depend on the 'rdf_extras' module, which is obsolete when
using RDFLib 4.0 and newer. This is good, because installing rdf-extras
version 0.4 from PyPI currently fails with this error:

    error: Installed distribution pyparsing 2.0.3 conflicts with
    requirement pyparsing<=1.5.7